refactor: extract concept set logic to useConceptSets composable
